Perform queries in batches to avoid running out of memory.  This is going to make the estimated time status messages less useful but I'll look at that again later.
git-svn-id: https://svn.internet2.edu/svn/i2mi/trunk@6521 bc24b658-cf39-4249-b827-e3526eaaeb42

+++ grouper/src/grouper/edu/internet2/middleware/grouper/misc/GrouperReport.java
@@ -187,7 +187,7 @@ public class GrouperReport {
       
       String syncFlatTablesResults = "Not configured to compute this today";
       String syncFlatTablesReport = null;
-      int syncFlatTablesCount = 0;
+      long syncFlatTablesCount = 0;
       if (syncFlatTables) {
         boolean saveUpdates = GrouperLoaderConfig.getPropertyBoolean("daily.report.syncFlatTables.saveUpdates", true);
         SyncFlatTables sync = new SyncFlatTables();
